Update instructions and links
Updated wording of instructions for saving etc. to match how the Jupyter notebook currently looks and works. Removed information that is not correct anymore (iris documentation doesn't open in a pop up window, Alt+ dash doesn't seem to work here for <- , at least not on a Mac when the same works in RStudio?). Fixed tidyverse link.